+++ src/pools/pool.ts
@@ -141,7 +141,7 @@ export interface IPool<
    */
   readonly emitter?: PoolEmitter
   /**
-   * Finds a free worker node key based on the number of tasks the worker has applied.
+   * Finds the first free worker node key based on the number of tasks the worker has applied.
    *
    * If a worker is found with `0` running tasks, it is detected as free and its worker node key is returned.
    *
@@ -150,6 +150,16 @@ export interface IPool<
    * @returns A worker node key if there is one, `-1` otherwise.
    */
   findFreeWorkerNodeKey: () => number
+  /**
+   * Finds the last free worker node key based on the number of tasks the worker has applied.
+   *
+   * If a worker is found with `0` running tasks, it is detected as free and its worker node key is returned.
+   *
+   * If no free worker is found, `-1` is returned.
+   *
+   * @returns A worker node key if there is one, `-1` otherwise.
+   */
+  findLastFreeWorkerNodeKey: () => number
   /**
    * Executes the function specified in the worker constructor with the task data input parameter.
    *