+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Optimize free worker finding in worker choice strategies.
+
 ## [2.4.10] - 2023-04-15
 
 ### Fixed